chore(rxMisc): decompose monolithic source code
