+++ src/rxMisc/scripts/rxAutoSave.spec.js
@@ -1,97 +1,5 @@
 /* jshint node: true */
 
-describe('rxDOMHelper', function () {
-    var rxjq;
-
-    var windowMock = {
-        getComputedStyle: sinon.stub()
-            .returns({ width: '10px', height: '20px' })
-    };
-
-    beforeEach(function () {
-        module('encore.ui.rxMisc');
-
-        module(function ($provide) {
-            $provide.value('$window', windowMock);
-        });
-
-        inject(function (rxDOMHelper) {
-            rxjq = rxDOMHelper;
-        });
-    });
-
-    it('should accept raw DOM objects', function () {
-        var div = angular.element('<div></div>'),
-            raw = div[0];
-
-        rxjq.width(raw);
-        expect(windowMock.getComputedStyle).to.have.been.calledWith(raw);
-    });
-
-    it('should accept jquery lite objects and convert them to raw DOM', function () {
-        var div = angular.element('<div></div>'),
-            raw = div[0];
-
-        rxjq.width(div);
-        expect(windowMock.getComputedStyle).to.have.been.calledWith(raw);
-    });
-
-    it('should allow query selectors for finding nested objects', function () {
-        var el = angular.element('<div><span><h1 class="my-title">Title!</h1></span></div>'),
-            title = rxjq.find(el, '.my-title');
-
-        expect(title.text()).to.equal('Title!');
-        expect(title[0].nodeName).to.equal('H1');
-    });
-
-    it('should wrap a single element', function () {
-        var span = $('<span class="hi">Test</span>');
-        var div = $('<div></div>');
-        rxjq.wrapAll(div[0], span[0]);
-        expect(div.find('span').text()).to.equal('Test');
-        expect(div.find('span').hasClass('hi')).to.be.true;
-    });
-
-    it('should wrap embedded elements', function () {
-        var span = $('<span class="hi"><span class="foo">Test</span></span>');
-        var div = $('<div></div>');
-        rxjq.wrapAll(div[0], span[0]);
-        expect(div.find('span.hi span.foo').text()).to.equal('Test');
-        expect(div.find('span.hi span.foo').hasClass('foo')).to.be.true;
-    });
-
-    it('should work for select elements', function () {
-        var select = $('<select><option class="foo">Foo</option><option class="bar">Bar</option></select>');
-        var div = $('<div></div>');
-        rxjq.wrapAll(div[0], select[0]);
-        expect(div.find('select > option.foo').text()).to.equal('Foo');
-        expect(div.find('select > option.foo').hasClass('foo')).to.be.true;
-        expect(div.find('select > option.bar').text()).to.equal('Bar');
-        expect(div.find('select > option.bar').hasClass('bar')).to.be.true;
-    });
-});
-
-describe('titleize', function () {
-    var titleize;
-
-    beforeEach(function () {
-        module('encore.ui.rxMisc');
-
-        inject(function (titleizeFilter) {
-            titleize = titleizeFilter;
-        });
-    });
-
-    it('replaces underscores with spaces', function () {
-        expect(titleize('_A_B_')).to.equal(' A B ');
-    });
-
-    it('converts the string to title case', function () {
-        expect(titleize('a bcD_e')).to.equal('A Bcd E');
-    });
-
-});
-
 describe('rxAutoSave', function () {
     var $rootScope, $q, $timeout, scope, rxAutoSave, a, b, LocalStorage, SessionStorage,
         now;
@@ -492,84 +400,5 @@ describe('rxAutoSave', function () {
 
         a.save();
         expect(a.getStoredValue().foo).to.equal('bar');
-    });
-});
-
-describe('rxNestedElement', function () {
-    var subject, result;
-
-    beforeEach(function () {
-        module('encore.ui.rxMisc');
-
-        inject(function (rxNestedElement) {
-            subject = rxNestedElement;
-        });
-    });
-
-    it('should return a function', function () {
-        expect(subject).to.be.a.function;
-    });
-
-    describe('result when called without options', function () {
-        beforeEach(function () {
-            result = subject();
-        });
-
-        it('should be an object', function () {
-            expect(result).to.be.an.object;
-        });
-
-        it('should define "controller"', function () {
-            expect(result.controller).to.be.defined;
-        });
-
-        it('should not define "link"', function () {
-            expect(result.link).to.not.be.defined;
-        });
-
-        it('should define "restrict"', function () {
-            expect(result.restrict).to.be.defined;
-        });
-
-        it('should restrict "E"', function () {
-            expect(result.restrict).to.eq('E');
-        });
-
-        it('should not define "require"', function () {
-            expect(result.require).to.not.be.defined;
-        });
-    });
-
-    describe('result when called with options', function () {
-        var opts = {
-            parent: 'foobar',
-            // yes, this isn't valid with angular
-            // but we're only testing if it overrides the default
-            restrict: 'Z'
-        };
-
-        beforeEach(function () {
-            result = subject(opts);
-        });
-
-        it('should restrict "Z"', function () {
-            expect(result.restrict).to.eq('Z');
-        });
-
-        it('should define "require"', function () {
-            expect(result.require).to.be.defined;
-        });
-
-        it('should have expected require', function () {
-            expect(result.require).to.eq('^foobar');
-        });
-
-        it('should define "link"', function () {
-            expect(result.link).to.be.defined;
-        });
-
-        it('should define "link" as a function', function () {
-            expect(result.link).to.be.a.function;
-        });
     });
 });
